Changed to log error msg about unknown link vs exiting with error. There is an existing broken link that the team will fix. We don't want the script to crash on this. It should keep processing. Also fixed link to homepage of tech docs
